FEATURE: Allow invalid plural keys
This patch introduces a new option, `strict:`. By default, its value is
`true` so if invalid plural keys are provided, an exception will be
raised (this is the current behavior).

Now, when calling `MessageFormat.compile` with `strict: false`, if there
are invalid plural keys (`many` for English, for example), no exception
will be thrown, and the JS code will be returned.